Update contact_me.php
I don't know php but line 31 ends in a semi colon from 32 through the end of the file ends in periods. Is that how it works? 

Also, line 45 has a period before the equal sign when declaring what I think is a variable. Not sure if that's how it works or not.